+++ harpoon/option_spec/harpoon_specs.py
@@ -232,6 +232,7 @@ class HarpoonSpec(object):
             , cache_from = delayed(or_spec(boolean(), listof(formatted(string_spec(), formatter=MergedOptionStringFormatter))))
             , squash_after = optional_spec(or_spec(boolean(), container_spec(Commands, listof(command_spec()))))
             , squash_before_push = optional_spec(or_spec(boolean(), container_spec(Commands, listof(command_spec()))))
+            , cleanup_intermediate_images = defaulted(boolean(), True)
             , persistence = optional_spec(create_spec(image_objs.Persistence
                 , validators.deprecated_key("persist", "Use ``folders``")
                 , action = required(formatted(string_spec(), formatter=MergedOptionStringFormatter))